factor handlers tests to their own module
