+++ src/handlers/mod.rs
@@ -1,5 +1,7 @@
 mod handler;
 mod serve;
+#[cfg(test)]
+mod tests;
 
 pub use handler::{Handler, Meta};
 pub use serve::serve;
